Ported vlan page to vuejs
